Fixed morton BVH builder to work with larger meshes. There still could be a potential issue with GPU texture object indexing limitations, but CPU tracing will be fine.

+++ src/raytracing/SplitBVH.h
@@ -30,7 +30,7 @@ SplitBVH(float * verts, int numPrimitives, int primitveType)
 	
 }
 
-void getFlatArray(int &size, int &leafSize, float *& inner, float *&leafs)
+void getFlatArray(int &size, int &leafSize, float *& inner, int *&leafs)
 {
 	return builder->bvhToFlatArray(root,size, leafSize, inner, leafs);
 }
